+++ src/promptflow/promptflow/_sdk/_submitter/utils.py
@@ -41,6 +41,7 @@ from promptflow._sdk._errors import InvalidFlowError, RunOperationError
 from promptflow._sdk._load_functions import load_flow
 from promptflow._sdk._utils import (
     _merge_local_code_and_additional_includes,
+    generate_flow_tools_json,
     get_local_connections_from_executable,
     get_used_connection_names_from_dict,
     update_dict_value_with_connections,
@@ -103,7 +104,7 @@ def overwrite_connections(flow_dag: dict, connections: dict, working_dir: PathLi
     executable_flow = ExecutableFlow._from_dict(flow_dag=flow_dag, working_dir=Path(working_dir))
 
     # generate tool meta for deployment name, model override
-    # tools_meta = generate_flow_tools_json(flow_directory=working_dir, dump=False, used_packages_only=True)
+    tools_meta = generate_flow_tools_json(flow_directory=working_dir, dump=False, used_packages_only=True)
 
     node_name_2_node = {node["name"]: node for node in flow_dag[NODES]}
 
@@ -126,7 +127,7 @@ def overwrite_connections(flow_dag: dict, connections: dict, working_dir: PathLi
             override_python_connections(
                 node=node,
                 connection_dict=connection_dict,
-                tools_meta={},
+                tools_meta=tools_meta,
                 executable_flow=executable_flow,
                 node_name=node_name,
             )