+++ src/promptflow/tests/test_configs/flows/connection_not_provided/print_env.py
@@ -1,0 +1,10 @@
+import os
+
+from promptflow import tool
+from promptflow.connections import CustomConnection
+
+
+@tool
+def get_env_var(key: str, connection: CustomConnection):
+    # get from env var
+    return {"value": os.environ.get(key)}